requirements.txt
openai
python-dotenv
discord.py
requests